+++ kernprof.py
@@ -51,9 +51,20 @@ which displays:
 
 .. code::
 
-    usage: kernprof [-h] [-V] [-l] [-b] [-o OUTFILE] [-s SETUP] [-v] [-r] [-u UNIT] [-z] [-i [OUTPUT_INTERVAL]] [-p PROF_MOD] [-m] [--prof-imports] {script | -m module} ...
+    usage: kernprof [-h] [-V] [-c CONFIG] [--no-config]
+                    [--line-by-line [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [--builtin [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [-s SETUP] [-p PROF_MOD]
+                    [--prof-imports [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [-o OUTFILE]
+                    [--view [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [--rich [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [-u UNIT]
+                    [--skip-zero [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]]
+                    [-i [OUTPUT_INTERVAL]]
+                    {script | -m module} ...
 
-    Run and profile a python script.
+    Run and profile a python script or module.
 
     positional arguments:
       {script | -m module}  The python script file or module to run
@@ -62,23 +73,59 @@ which displays:
     options:
       -h, --help            show this help message and exit
       -V, --version         show program's version number and exit
-      -l, --line-by-line    Use the line-by-line profiler instead of cProfile. Implies --builtin.
-      -b, --builtin         Put 'profile' in the builtins. Use 'profile.enable()'/'.disable()', '@profile' to decorate functions, or 'with profile:' to profile a section of code.
-      -o OUTFILE, --outfile OUTFILE
-                            Save stats to <outfile> (default: 'scriptname.lprof' with --line-by-line, 'scriptname.prof' without)
-      -s SETUP, --setup SETUP
-                            Code to execute before the code to profile
-      -v, --view            View the results of the profile in addition to saving it
-      -r, --rich            Use rich formatting if viewing output
-      -u UNIT, --unit UNIT  Output unit (in seconds) in which the timing info is displayed (default: 1e-6)
-      -z, --skip-zero       Hide functions which have not been called
-      -i [OUTPUT_INTERVAL], --output-interval [OUTPUT_INTERVAL]
-                            Enables outputting of cumulative profiling results to file every n seconds. Uses the threading module. Minimum value is 1 (second). Defaults to disabled.
+      -c, --config CONFIG   Path to the TOML file, from the `tool.line_profiler.kernprof`
+                            table of which to load defaults for the options. (Default:
+                            'pyproject.toml')
+      --no-config           Disable the loading of configuration files other than the
+                            default one
+
+    profiling options:
+      --line-by-line [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            Use the line-by-line profiler instead of cProfile. Implies
+                            `--builtin`. (Default: False; short form: -l)
+      --builtin [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            Put `profile` in the builtins. Use
+                            `profile.enable()`/`.disable()` to toggle profiling,
+                            `@profile` to decorate functions, or `with profile:` to
+                            profile a section of code. (Default: False; short form: -b)
+      -s, --setup SETUP     Path to the Python source file containing setup code to
+                            execute before the code to profile. (Default: N/A)
       -p, --prof-mod PROF_MOD
-                            List of modules, functions and/or classes to profile specified by their name or path. List is comma separated, adding the current script path profiles
-                            the full script. Multiple copies of this flag can be supplied and the.list is extended. Only works with line_profiler -l, --line-by-line
-      --prof-imports        If specified, modules specified to `--prof-mod` will also autoprofile modules that they import. Only works with line_profiler -l, --line-by-line
-"""
+                            List of modules, functions and/or classes to profile specified
+                            by their name or path. List is comma separated, adding the
+                            current script path profiles the full script. Multiple copies
+                            of this flag can be supplied and the list is extended (e.g.
+                            `-p this.module,another.module -p some.func`). Only works with
+                            line profiling (`-l`/`--line-by-line`). (Default: N/A)
+      --prof-imports [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            If the script/module profiled is in `--prof-mod`, autoprofile
+                            all its imports. Only works with line profiling (`-l`/`--line-
+                            by-line`). (Default: False)
+
+    output options:
+      -o, --outfile OUTFILE
+                            Save stats to OUTFILE. (Default:
+                            '<script_or_module_name>.lprof' in line-profiling mode
+                            (`-l`/`--line-by-line`); '<script_or_module_name>.prof'
+                            otherwise)
+      --view [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            View the results of the profile in addition to saving it.
+                            (Default: False; short form: -v)
+      --rich [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            Use rich formatting if viewing output. (Default: False; short
+                            form: -r)
+      -u, --unit UNIT       Output unit (in seconds) in which the timing info is
+                            displayed. (Default: 1e-06 s)
+      --skip-zero [Y[es] | N[o] | T[rue] | F[alse] | on | off | 1 | 0]
+                            Hide functions which have not been called. (Default: False;
+                            short form: -z)
+      -i, --output-interval [OUTPUT_INTERVAL]
+                            Enables outputting of cumulative profiling results to OUTFILE
+                            every OUTPUT_INTERVAL seconds. Uses the threading module.
+                            Minimum value (and the value implied if the bare option is
+                            given) is 1 s. (Default: 0 s (disabled))
+"""  # noqa: E501
+import argparse
 import builtins
 import functools
 import os
@@ -87,7 +134,6 @@ import threading
 import asyncio  # NOQA
 import concurrent.futures  # NOQA
 import time
-from argparse import ArgumentError, ArgumentParser
 from runpy import run_module
 
 # NOTE: This version needs to be manually maintained in
@@ -101,6 +147,11 @@ try:
 except ImportError:
     from profile import Profile  # type: ignore[assignment,no-redef]
 
+import line_profiler
+from line_profiler.cli_utils import (
+    add_argument, get_cli_config,
+    get_python_executable as _python_command,  # Compatibility
+    positive_float, short_string_path)
 from line_profiler.profiler_mixin import ByCountProfilerMixin
 
 
@@ -151,7 +202,7 @@ class RepeatedTimer:
 
     References:
         .. [SO474528] https://stackoverflow.com/questions/474528/execute-function-every-x-seconds/40965385#40965385
-    """
+    """  # noqa: E501
     def __init__(self, interval, dump_func, outfile):
         self._timer = None
         self.interval = interval
@@ -169,7 +220,8 @@ class RepeatedTimer:
     def start(self):
         if not self.is_running:
             self.next_call += self.interval
-            self._timer = threading.Timer(self.next_call - time.time(), self._run)
+            self._timer = threading.Timer(self.next_call - time.time(),
+                                          self._run)
             self._timer.start()
             self.is_running = True
 
@@ -178,7 +230,7 @@ class RepeatedTimer:
         self.is_running = False
 
 
-def find_module_script(module_name):
+def find_module_script(module_name, *, exit_on_error=True):
     """Find the path to the executable script for a module or package."""
     from line_profiler.autoprofile.util_static import modname_to_modpath
 
@@ -187,11 +239,15 @@ def find_module_script(module_name):
         if fname:
             return fname
 
-    sys.stderr.write('Could not find module %s\n' % module_name)
-    raise SystemExit(1)
+    msg = f'Could not find module `{module_name}`'
+    if exit_on_error:
+        print(msg, file=sys.stderr)
+        raise SystemExit(1)
+    else:
+        raise ModuleNotFoundError(msg)
 
 
-def find_script(script_name):
+def find_script(script_name, *, exit_on_error=True):
     """ Find the script.
 
     If the input is not a file, then $PATH will be searched.
@@ -206,21 +262,12 @@ def find_script(script_name):
         if os.path.isfile(fn):
             return fn
 
-    sys.stderr.write('Could not find script %s\n' % script_name)
-    raise SystemExit(1)
-
-
-def _python_command():
-    """
-    Return a command that corresponds to :py:obj:`sys.executable`.
-    """
-    import shutil
-    if shutil.which('python') == sys.executable:
-        return 'python'
-    elif shutil.which('python3') == sys.executable:
-        return 'python3'
+    msg = f'Could not find script {script_name!r}'
+    if exit_on_error:
+        print(msg, file=sys.stderr)
+        raise SystemExit(1)
     else:
-        return sys.executable
+        raise FileNotFoundError(msg)
 
 
 class _restore_list:
@@ -328,19 +375,15 @@ def pre_parse_single_arg_directive(args, flag, sep='--'):
 
 @_restore_list(sys.argv)
 @_restore_list(sys.path)
-def main(args=None):
+def main(args=None, exit_on_error=True):
     """
     Runs the command line interface
     """
-    def positive_float(value):
-        val = float(value)
-        if val <= 0:
-            raise ArgumentError
-        return val
-
     create_parser = functools.partial(
-        ArgumentParser,
-        description='Run and profile a python script.')
+        argparse.ArgumentParser,
+        description='Run and profile a python script or module.')
+    get_kernprof_config = functools.partial(get_cli_config, 'kernprof')
+    defaults, default_source = get_kernprof_config()
 
     if args is None:
         args = sys.argv[1:]
@@ -363,55 +406,134 @@ def main(args=None):
         help_parser = create_parser()
         parsers = [real_parser, help_parser]
     for parser in parsers:
-        parser.add_argument('-V', '--version', action='version', version=__version__)
-        parser.add_argument('-l', '--line-by-line', action='store_true',
-                            help='Use the line-by-line profiler instead of cProfile. Implies --builtin.')
-        parser.add_argument('-b', '--builtin', action='store_true',
-                            help="Put 'profile' in the builtins. Use 'profile.enable()'/'.disable()', "
-                            "'@profile' to decorate functions, or 'with profile:' to profile a "
-                            'section of code.')
-        parser.add_argument('-o', '--outfile',
-                            help="Save stats to <outfile> (default: 'scriptname.lprof' with "
-                            "--line-by-line, 'scriptname.prof' without)")
-        parser.add_argument('-s', '--setup',
-                            help='Code to execute before the code to profile')
-        parser.add_argument('-v', '--view', action='store_true',
-                            help='View the results of the profile in addition to saving it')
-        parser.add_argument('-r', '--rich', action='store_true',
-                            help='Use rich formatting if viewing output')
-        parser.add_argument('-u', '--unit', default='1e-6', type=positive_float,
-                            help='Output unit (in seconds) in which the timing info is '
-                            'displayed (default: 1e-6)')
-        parser.add_argument('-z', '--skip-zero', action='store_true',
-                            help="Hide functions which have not been called")
-        parser.add_argument('-i', '--output-interval', type=int, default=0, const=0, nargs='?',
-                            help="Enables outputting of cumulative profiling results to file every n seconds. Uses the threading module. "
-                            "Minimum value is 1 (second). Defaults to disabled.")
-        parser.add_argument('-p', '--prof-mod', action='append', type=str,
-                            help="List of modules, functions and/or classes to profile specified by their name or path. "
-                            "List is comma separated, adding the current script path profiles the full script. "
-                            "Multiple copies of this flag can be supplied and the.list is extended. "
-                            "Only works with line_profiler -l, --line-by-line")
-        parser.add_argument('--prof-imports', action='store_true',
-                            help="If specified, modules specified to `--prof-mod` will also autoprofile modules that they import. "
-                            "Only works with line_profiler -l, --line-by-line")
+        add_argument(parser, '-V', '--version',
+                     action='version', version=__version__)
+        add_argument(parser, '-c', '--config',
+                     help='Path to the TOML file, from the '
+                     '`tool.line_profiler.kernprof` table of which to load '
+                     'defaults for the options. '
+                     f'(Default: {short_string_path(default_source)!r})')
+        add_argument(parser, '--no-config',
+                     action='store_const', dest='config', const=False,
+                     help='Disable the loading of configuration files other '
+                     'than the default one')
+        prof_opts = parser.add_argument_group('profiling options')
+        add_argument(prof_opts, '-l', '--line-by-line', action='store_true',
+                     help='Use the line-by-line profiler instead of cProfile. '
+                     'Implies `--builtin`. '
+                     f'(Default: {defaults["line_by_line"]})')
+        add_argument(prof_opts, '-b', '--builtin', action='store_true',
+                     help="Put `profile` in the builtins. "
+                     "Use `profile.enable()`/`.disable()` to "
+                     "toggle profiling, "
+                     "`@profile` to decorate functions, "
+                     "or `with profile:` to profile a section of code. "
+                     f"(Default: {defaults['builtin']})")
+        if defaults['setup']:
+            def_setupfile = repr(defaults['setup'])
+        else:
+            def_setupfile = 'N/A'
+        add_argument(prof_opts, '-s', '--setup',
+                     help='Path to the Python source file containing setup '
+                     'code to execute before the code to profile. '
+                     f'(Default: {def_setupfile})')
+        if defaults['prof_mod']:
+            def_prof_mod = repr(defaults['prof_mod'])
+        else:
+            def_prof_mod = 'N/A'
+        add_argument(prof_opts, '-p', '--prof-mod', action='append',
+                     help="List of modules, functions and/or classes "
+                     "to profile specified by their name or path. "
+                     "List is comma separated, adding the current script path "
+                     "profiles the full script. "
+                     "Multiple copies of this flag can be supplied and "
+                     "the list is extended "
+                     "(e.g. `-p this.module,another.module -p some.func`). "
+                     "Only works with line profiling (`-l`/`--line-by-line`). "
+                     f"(Default: {def_prof_mod})")
+        add_argument(prof_opts, '--prof-imports', action='store_true',
+                     help="If the script/module profiled is in `--prof-mod`, "
+                     "autoprofile all its imports. "
+                     "Only works with line profiling (`-l`/`--line-by-line`). "
+                     f"(Default: {defaults['prof_imports']})")
+        out_opts = parser.add_argument_group('output options')
+        if defaults['outfile']:
+            def_outfile = repr(defaults['outfile'])
+        else:
+            def_outfile = (
+                "'<script_or_module_name>.lprof' in line-profiling mode "
+                "(`-l`/`--line-by-line`); "
+                "'<script_or_module_name>.prof' otherwise")
+        add_argument(out_opts, '-o', '--outfile',
+                     help=f'Save stats to OUTFILE. (Default: {def_outfile})')
+        add_argument(out_opts, '-v', '--view', action='store_true',
+                     help='View the results of the profile '
+                     'in addition to saving it. '
+                     f'(Default: {defaults["view"]})')
+        add_argument(out_opts, '-r', '--rich', action='store_true',
+                     help='Use rich formatting if viewing output. '
+                     f'(Default: {defaults["rich"]})')
+        add_argument(out_opts, '-u', '--unit', type=positive_float,
+                     help='Output unit (in seconds) in which '
+                     'the timing info is displayed. '
+                     f'(Default: {defaults["unit"]} s)')
+        add_argument(out_opts, '-z', '--skip-zero', action='store_true',
+                     help="Hide functions which have not been called. "
+                     f"(Default: {defaults['skip_zero']})")
+        if defaults['output_interval']:
+            def_out_int = f'{defaults["output_interval"]} s'
+        else:
+            def_out_int = '0 s (disabled)'
+        add_argument(out_opts, '-i', '--output-interval',
+                     type=int, const=1, nargs='?',
+                     help="Enables outputting of cumulative profiling results "
+                     "to OUTFILE every OUTPUT_INTERVAL seconds. "
+                     "Uses the threading module. "
+                     "Minimum value (and the value implied if the bare option "
+                     f"is given) is 1 s. (Default: {def_out_int})")
 
         if parser is help_parser or module is None:
-            parser.add_argument('script',
-                                metavar='{script | -m module}',
-                                help='The python script file or module to run')
-        parser.add_argument('args', nargs='...', help='Optional script arguments')
+            add_argument(parser, 'script',
+                         metavar='{script | -m module}',
+                         help='The python script file or module to run')
+        add_argument(parser, 'args',
+                     nargs='...', help='Optional script arguments')
 
     # Hand off to the dummy parser if necessary to generate the help
     # text
-    options = real_parser.parse_args(args)
+    try:
+        options = real_parser.parse_args(args)
+    except SystemExit as e:
+        # If `exit_on_error` is true, let `SystemExit` bubble up and
+        # kill the interpretor;
+        # else, catch and handle it more gracefully
+        # (Note: can't use `ArgumentParser(exit_on_error=False)` in
+        # Python 3.8)
+        if exit_on_error:
+            raise
+        elif e.code:
+            raise RuntimeError from None
+        else:
+            return
     if help_parser and getattr(options, 'help', False):
-        # This should raise a `SystemExit`
-        help_parser.parse_args([*args, '-m', module])
+        help_parser.print_help()
+        if exit_on_error:
+            raise SystemExit(0)
+        else:
+            return
+
+    # Parse the provided config file (if any), and resolve the values
+    # of the un-specified options
     try:
         del options.help
     except AttributeError:
         pass
+    if options.config:
+        defaults, options.config = get_kernprof_config(options.config)
+    for key, default in defaults.items():
+        if getattr(options, key, None) is None:
+            setattr(options, key, default)
+
     # Add in the pre-partitioned arguments cut off by `-m <module>`
     options.args += post_args
     if module is not None:
@@ -419,7 +541,7 @@ def main(args=None):
 
     if not options.outfile:
         extension = 'lprof' if options.line_by_line else 'prof'
-        options.outfile = '%s.%s' % (os.path.basename(options.script), extension)
+        options.outfile = f'{os.path.basename(options.script)}.{extension}'
 
     sys.argv = [options.script] + options.args
     if module:
@@ -428,10 +550,10 @@ def main(args=None):
         # Note: this NEEDS to happen here, before the setup script (or
         # any other code) has a chance to `os.chdir()`
         sys.path.insert(0, os.path.abspath(os.curdir))
-    if options.setup is not None:
-        # Run some setup code outside of the profiler. This is good for large
-        # imports.
-        setup_file = find_script(options.setup)
+    if options.setup:
+        # Run some setup code outside of the profiler. This is good for
+        # large imports.
+        setup_file = find_script(options.setup, exit_on_error=exit_on_error)
         __file__ = setup_file
         __name__ = '__main__'
         # Make sure the script's directory is on sys.path instead of just
@@ -441,7 +563,6 @@ def main(args=None):
         execfile(setup_file, ns, ns)
 
     if options.line_by_line:
-        import line_profiler
         prof = line_profiler.LineProfiler()
         options.builtin = True
     elif Profile.__module__ == 'profile':
@@ -450,36 +571,29 @@ def main(args=None):
     else:
         prof = ContextualProfile()
 
-    # If line_profiler is installed, then overwrite the explicit decorator
-    try:
-        import line_profiler
-    except ImportError:  # Shouldn't happen
-        install_profiler = global_profiler = None
-    else:
-        global_profiler = line_profiler.profile
-        install_profiler = global_profiler._kernprof_overwrite
-
-    if global_profiler:
-        install_profiler(prof)
+    # Overwrite the explicit decorator
+    global_profiler = line_profiler.profile
+    install_profiler = global_profiler._kernprof_overwrite
+    install_profiler(prof)
 
     if options.builtin:
         builtins.__dict__['profile'] = prof
 
     if module:
-        script_file = find_module_script(options.script)
+        script_file = find_module_script(options.script,
+                                         exit_on_error=exit_on_error)
     else:
-        script_file = find_script(options.script)
+        script_file = find_script(options.script, exit_on_error=exit_on_error)
         # Make sure the script's directory is on sys.path instead of
         # just kernprof.py's.
         sys.path.insert(0, os.path.dirname(script_file))
     __file__ = script_file
     __name__ = '__main__'
 
-    if options.output_interval:
-        rt = RepeatedTimer(max(options.output_interval, 1), prof.dump_stats, options.outfile)
     original_stdout = sys.stdout
     if options.output_interval:
-        rt = RepeatedTimer(max(options.output_interval, 1), prof.dump_stats, options.outfile)
+        rt = RepeatedTimer(max(options.output_interval, 1), prof.dump_stats,
+                           options.outfile)
     try:
         try:
             execfile_ = execfile
@@ -500,21 +614,24 @@ def main(args=None):
                                 prof_mod=prof_mod,
                                 profile_imports=options.prof_imports,
                                 as_module=module is not None)
-            elif module and options.builtin:
-                rmod_(options.script, ns)
-            elif options.builtin:
-                execfile(script_file, ns, ns)
-            elif module:
-                prof.runctx(f'rmod_({options.script!r}, globals())', ns, ns)
             else:
-                prof.runctx('execfile_(%r, globals())' % (script_file,), ns, ns)
+                if module:
+                    runner, target = 'rmod_', options.script
+                else:
+                    runner, target = 'execfile_', script_file
+                assert runner in ns
+                if options.builtin:
+                    ns[runner](target, ns)
+                else:
+                    prof.runctx(f'{runner}({target!r}, globals())', ns, ns)
         except (KeyboardInterrupt, SystemExit):
             pass
     finally:
         if options.output_interval:
             rt.stop()
         prof.dump_stats(options.outfile)
-        print('Wrote profile results to %s' % options.outfile)
+        short_outfile = short_string_path(options.outfile)
+        print(f'Wrote profile results to {short_outfile!r}')
         if options.view:
             if isinstance(prof, ContextualProfile):
                 prof.print_stats()
@@ -522,18 +639,18 @@ def main(args=None):
                 prof.print_stats(output_unit=options.unit,
                                  stripzeros=options.skip_zero,
                                  rich=options.rich,
-                                 stream=original_stdout)
+                                 stream=original_stdout,
+                                 config=options.config)
         else:
             print('Inspect results with:')
             py_exe = _python_command()
             if isinstance(prof, ContextualProfile):
-                print(f'{py_exe} -m pstats "{options.outfile}"')
+                print(f'{py_exe} -m pstats {short_outfile!r}')
             else:
-                print(f'{py_exe} -m line_profiler -rmt "{options.outfile}"')
+                print(f'{py_exe} -m line_profiler -rmt {short_outfile!r}')
         # Restore the state of the global `@line_profiler.profile`
-        if global_profiler:
-            install_profiler(None)
+        install_profiler(None)
 
 
 if __name__ == '__main__':
-    main(sys.argv[1:])
+    main()